requirements.txt
argparse==1.3.0
Babel==1.3
backports.ssl_match_hostname==3.4.0.2
bitmath==1.2.3-4
boto==2.38.0
boto3==1.2.2
botocore==1.3.9
cffi==1.1.2
characteristic==14.3.0
cryptography==1.1.2
debtcollector==0.5.0
docker-py==1.6.0
effect==0.10
eliot==0.11.0
eliot-tree==15.3.0
enum34==1.0.4
futures==2.2.0
hypothesis==1.14.0
idna==2.0
ipaddr==2.1.11
ipaddress==1.0.7
iso8601==0.1.10
jmespath==0.9.0
jsonschema==2.4.0
klein==0.2.3
machinist==0.2.0
msgpack-python==0.4.6
ndg-httpsclient==0.4.0
netaddr==0.7.14
netifaces==0.10.4
oslo.config==1.12.1
oslo.i18n==1.7.0
oslo.serialization==1.6.0
oslo.utils==1.6.0
pbr==0.11.0
pip==7.1.0
prettytable==0.7.2
psutil==2.1.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==0.15.1
pyrsistent==0.11.9
python-cinderclient==1.1.1
python-keystoneclient==1.4.0
python-keystoneclient-rackspace==0.1.3
python-novaclient==2.24.1
pytz==2015.4
PyYAML==3.10
repoze.lru==0.6
requests==2.7.0
service-identity==14.0.0
setuptools==18.0.1
simplejson==3.7.3
six==1.9.0
stevedore==1.5.0
toolz==0.7.4
treq==0.2.1
Twisted==15.5.0
websocket-client==0.32.0
Werkzeug==0.10.4
wheel==0.24.0
wrapt==1.10.4
zope.interface==4.1.2
txeffect==0.9
attrs==15.1.0
